+++ CHANGELOG.md
@@ -1,4 +1,14 @@
-* 修复批量添加账号的问题,支持批量添加tiktok和instagram账号
-* Fix the problem of batch adding accounts, support batch adding tiktok and instagram accounts
-* 优化复制粘贴功能
-* Optimize the copy and paste function
+# Change Log
+
+- Add more network options for users to choose from
+
+## Improvements
+
+- Support more network for usdt payment
+  - Arbitrum One
+  - Avalanche C-Chain
+  - Optimism
+  - Polygon
+  - Solana
+  - TON
+  - X Layer